% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adm_ada.R
\docType{data}
\name{grade_levels}
\alias{grade_levels}
\title{The factor levels that correspond to grades in RUSD}
\format{
An object of class \code{character} of length 17.
}
\usage{
grade_levels
}
\description{
The factor levels that correspond to grades in RUSD
}
\keyword{datasets}
