% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_grade_level.R
\name{parse_grade_level}
\alias{parse_grade_level}
\title{Title}
\usage{
parse_grade_level(.column, pad_chars = c("\\\\s", 0))
}
\arguments{
\item{.column}{input data like c("1", "02", " 03", "KG")}

\item{pad_chars}{a vector of characters to remove}
}
\value{
a text vector with stripped grade numbers
}
\description{
Title
}
\examples{
parse_grade_level(c(1:3, "04", " 05", "PK", "KG"))
}
