% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_adm_ada.R
\name{read_adm_ada}
\alias{read_adm_ada}
\title{Read a CSV file with ADM and ADA attendance data}
\usage{
read_adm_ada(file_name)
}
\arguments{
\item{file_name}{The location of the file}
}
\value{
a tibble
}
\description{
Read a CSV file with ADM and ADA attendance data
}
